requirements.txt
altair==4.2.0
hiplot==0.1.33
matplotlib==3.5.1
missingno==0.5.1
pandas==1.4.2
Pillow==9.2.0
plotly==5.6.0
seaborn==0.11.2
streamlit==1.13.0
